komtek/requirements.txt
django
djangorestframework
django-filter
django-crispy-forms
django-bootstrap3
drf-yasg
